task 7 requires that the entered numbers are integers

+++ week-2/solutions/_7.c
@@ -1,10 +1,11 @@
 #include <stdio.h>
 
 int main(void) {
-    float a, b, c;
-    scanf("%f %f %f", &a, &b, &c);
+    int a, b, c;
+    scanf("%d %d %d", &a, &b, &c);
     
-    double answer = (a+b+c) / 3;
+    double answer = (a+b+c) / 3.0; // note that (a+b+c) / 3 will give an integer result
+   								   // despite storing the result in a variable of type double
 
     printf("%.2lf\n", answer);
